Split 6 FT
