+++ src/test/TaskTest.java
@@ -1,4 +1,4 @@
-package tests;
+package test;
 
 import interfaces.TaskManager;
 import manager.Managers;
